Updated constants file to include initial coordinates
